Emisor and RegimenFiscal Refactoring.
The xsd tool creates an Item property inside the RegimenFiscal class. This property is not used by the serializer. The only requirement is that the RegimenFiscal property is defined as an array of RegimenFiscal or as a List<RegimenFiscal> and the attribute of the Xml must be XmlElement instead of XmlArray. (There is no container for the collection, the RegimenFiscal is an unbound element that can be inside of the Emisor element).